All memory objects are now derived from the class Object. Please Read extended explination.
To allow a generic method of hooking events on memory objects, all
memory objects are now derived from a single base. This base "Object"
allows to set a event listener object which will receive init, destroy
and other events depending on the object. Most event triggering is not
yet implemented but the base is now set for future functionality.